automatically backed up by learn

+++ lib/cash_register.rb
@@ -45,7 +45,7 @@ class CashRegister
   end
 
   def void_last_transaction
-    @total = @total.to_i - @@total_cart[-1].to_i
+    @total = @total.to_i - @@total_cart[-1].to_f
   end
 
 end
